@@ -2,6 +2,7 @@
   "dependencies": {
     "@aws-sdk/client-secrets-manager": "^3.359.0",
     "@stripe/stripe-js": "^1.46.0",
+    "@supabase/ssr": "^0.0.10",
     "@supabase/supabase-js": "^2.15.0",
     "@tailwindcss/forms": "^0.5.3",
     "@types/sshpk": "^1.17.2",
@@ -11,13 +12,14 @@
     "eslint-config-next": "^13.2.4",
     "friendly-words": "^1.2.0",
     "generate-password": "^1.7.0",
-    "next": "^13.2.4",
+    "next": "^14.0.4",
     "pg": "^8.9.0",
     "pino": "^8.11.0",
     "raw-body": "^2.5.2",
     "react": "^18.2.0",
     "react-chartjs-2": "^5.2.0",
     "react-dom": "^18.2.0",
+    "sharp": "^0.33.0",
     "stripe": "^11.10.0",
     "stripe-event-types": "^2.3.0",
     "tweetnacl": "^1.0.3",
@@ -30,6 +32,7 @@
     "lint": "next lint"
   },
   "devDependencies": {
+    "@stylistic/eslint-plugin": "^1.5.0",
     "@types/aws-lambda": "^8.10.115",
     "@types/friendly-words": "^1.2.0",
     "@types/node": "^18.0.3",
@@ -37,7 +40,6 @@
     "@types/react": "^18.0.15",
     "@types/webfontloader": "^1.6.34",
     "autoprefixer": "^10.4.13",
-    "aws-cdk-lib": "2.79.1",
     "constructs": "10.1.156",
     "postcss": "^8.4.21",
     "sst": "^2.7.2",
