@@ -1,45 +1,34 @@
-import { FunctionComponent, useEffect, useState } from 'react'
-import { useRouter } from 'next/router'
+"use client";
+
+import { FunctionComponent } from 'react';
+import {usePathname} from 'next/navigation';
+import Image from "next/image";
+
+import kestrelTransparent from "../images/kestrel_transparent.png";
+import profileIconWhite from "../images/profile-icon-white.png";
+import twitterIcon from "../images/twitter-icon.svg";
+import githubLight from "../images/GitHub-Mark-Light-32px.png";
 
 type NavigationProps = {
-}
+};
 
 export const Navigation: FunctionComponent<NavigationProps> = () => {
-    const router = useRouter()
-    const defaultClasses = "nav-link w-nav-link"
-    const profileClasses = "nav-link w-inline-block social-icons"
-    const socialIcons = profileClasses + " w-hidden-small w-hidden-tiny"
-    const currentTabAppend = " w--current"
-    const [docsClasses, setDocsClasses] = useState(defaultClasses)
-    const [appsClasses, setAppsClasses] = useState(defaultClasses)
-    useEffect(
-        () => {
-            if (!router.isReady) return;
-            console.log(`Pathname: ${router.pathname}`)
-            switch (router.pathname) {
-                case "/docs":
-                    setDocsClasses(defaultClasses + currentTabAppend)
-                    break
-                case "/apps":
-                    setAppsClasses(defaultClasses + currentTabAppend)
-                    break
-                // case "/contact":
-                //     setContactClasses(defaultClasses + currentTabAppend)
-                //     break
-            }
-        },
-        [router.isReady],
-    )
+    const pathname = usePathname();
+    const defaultClasses = "nav-link w-nav-link";
+    const profileClasses = "nav-link w-inline-block social-icons";
+    const socialIcons = profileClasses + " w-hidden-small w-hidden-tiny";
+    const currentTabAppend = " w--current";
+
+    const docsClasses = pathname === "/docs" ? defaultClasses + currentTabAppend : defaultClasses;
+    const appsClasses = pathname === "/apps" ? defaultClasses + currentTabAppend : defaultClasses;
     return <div role="banner" className="navbar w-nav">
         <div className="w-container">
             <a href="/" className="brand w-clearfix w-nav-brand w--current">
-                <img
-                    src="/images/kestrel_transparent.png"
-                    width="75"
-                    srcSet="/images/kestrel_transparent-p-500.png 500w, /images/kestrel_transparent.png 508w"
-                    sizes="75px"
-                    alt=""
+                <Image
+                    src={kestrelTransparent}
+                    alt="Kestrel logo"
                     className="stone-logo"
+                    style={{width: 75, height: 30}}
                 />
                 <div className="logo-text w-hidden-tiny w-hidden-small">Kestrel</div>
             </a>
@@ -58,26 +47,26 @@ export const Navigation: FunctionComponent<NavigationProps> = () => {
                 className={profileClasses}
             >
                 Profile&nbsp;
-                <img src="/images/profile-icon-white.png" width="15" alt="" />
+                <Image src={profileIconWhite} style={{width: 15, height: 15}} alt="Profile icon" />
             </a>
             <a
                 href="https://www.twitter.com/substructureone"
                 className={socialIcons}
                 target="_blank"
             >
-                <img src="/images/twitter-icon.svg" width="16" alt="" />
+                <Image src={twitterIcon} style={{width: 16, height: 16}} alt="Twitter icon" />
             </a>
             <a
                 href="https://github.com/SubstructureOne"
                 className={socialIcons}
                 target="_blank"
             >
-                <img src="/images/GitHub-Mark-Light-32px.png" width="15" alt="" />
+                <Image src={githubLight} style={{width: 15, height: 15}} alt="Github icon" />
             </a>
             <div className="menu-button w-nav-button">
                 <div className="menu-icon w-icon-nav-menu"></div>
             </div>
         </div>
-    </div>
-}
+    </div>;
+};
 
