@@ -1,13 +1,15 @@
-import {NextPage} from "next"
-import {Headers} from "../components/Headers"
-import {Navigation} from "../components/Navigation"
-import Footer from "../components/Footer"
-import {useEffect, useRef} from "react"
-import {useSession} from "../utils/supabaseClient"
-import {AllTransactions} from "../utils/dbtypes"
+"use client";
+
+import {NextPage} from "next";
+import {Headers} from "../../components/Headers";
+import {Navigation} from "../../components/Navigation";
+import Footer from "../../components/Footer";
+import {useEffect, useRef} from "react";
+import {AllTransactions} from "../../utils/dbtypes";
+import {useSession} from "../../utils/supabase/client";
 
 
-type Callback = () => void
+type Callback = () => void;
 
 function useInterval(callback: Callback, delay: number) {
     const savedCallback = useRef<Callback>();
@@ -17,7 +19,7 @@ function useInterval(callback: Callback, delay: number) {
     useEffect(() => {
         function tick() {
             if (savedCallback.current !== undefined) {
-                savedCallback.current()
+                savedCallback.current();
             }
         }
         let id = setInterval(tick, delay);
@@ -25,9 +27,9 @@ function useInterval(callback: Callback, delay: number) {
     }, [delay]);
 }
 const ConfirmPayment: NextPage = () => {
-    const [session] = useSession()
-    let moreRecentThan = new Date()
-    moreRecentThan.setHours(moreRecentThan.getHours() - 1)
+    const [session] = useSession();
+    let moreRecentThan = new Date();
+    moreRecentThan.setHours(moreRecentThan.getHours() - 1);
     useInterval(() => {
         if (session !== null && session.user !== null) {
             fetch(
@@ -41,11 +43,11 @@ const ConfirmPayment: NextPage = () => {
                 .then(response => response.json())
                 .then(transactions => {
                     if ((transactions as AllTransactions).external_txns.filter(txn => new Date(txn.exttransaction_time) > moreRecentThan).length > 0) {
-                        (window as Window).location = "/profile"
+                        (window as Window).location = "/profile";
                     }
-                })
+                });
         }
-    }, 1000)
+    }, 1000);
     return <>
         <Headers title="Kestrel Substructure: Confirm Payment"/>
         <Navigation/>
@@ -55,7 +57,7 @@ const ConfirmPayment: NextPage = () => {
             <div className="my-10">Confirming payment...</div>
         </div>
         <Footer/>
-    </>
-}
+    </>;
+};
 
-export default ConfirmPayment
+export default ConfirmPayment;
