@@ -1,7 +1,8 @@
-import {NextPage} from "next"
-import {Headers} from "../components/Headers"
-import {Navigation} from "../components/Navigation"
-import Footer from "../components/Footer"
+import {NextPage} from "next";
+
+import {Headers} from "../../components/Headers";
+import {Navigation} from "../../components/Navigation";
+import Footer from "../../components/Footer";
 
 const PrivacyPolicy: NextPage = () => {
     return <>
@@ -140,7 +141,7 @@ const PrivacyPolicy: NextPage = () => {
             </p>
         </div>
         <Footer/>
-    </>
-}
+    </>;
+};
 
-export default PrivacyPolicy
+export default PrivacyPolicy;
