@@ -28,7 +28,7 @@ const Footer = () => {
                 </div>
             </div>
         </div>
-    </div>
-}
+    </div>;
+};
 
-export default Footer
+export default Footer;
