@@ -1,4 +1,4 @@
-import { Client } from "pg"
+import { Client } from "pg";
 import {
     ChargeInfo,
     AccountInfo,
@@ -6,16 +6,16 @@ import {
     NewExternalTransactionInfo,
     NewUserInfo,
     ExternalTransactionInfo
-} from "./dbtypes"
-import {getEnviron} from "./secrets"
-import {KResult, Ok, Err} from "./errors"
-import logger from "./logger"
-import {decryptDataWithKey, encryptDataWithKey} from "./encrypt_server"
+} from "./dbtypes";
+import {getEnviron} from "./secrets";
+import {KResult, Ok, Err} from "./errors";
+import logger from "./logger";
+import {decryptDataWithKey, encryptDataWithKey} from "./encrypt_server";
 
 export async function pgconnect(): Promise<KResult<Client>> {
-    const environ = await getEnviron()
+    const environ = await getEnviron();
     if (environ.isErr) {
-        return Err(environ.error)
+        return Err(environ.error);
     }
     const client = new Client({
         "host": environ.value.POSTGRES_HOST,
@@ -23,9 +23,9 @@ export async function pgconnect(): Promise<KResult<Client>> {
         "user": environ.value.POSTGRES_USER,
         "password": environ.value.POSTGRES_PASSWORD,
         "database": environ.value.POSTGRES_DATABASE,
-    })
-    await client.connect()
-    return Ok(client)
+    });
+    await client.connect();
+    return Ok(client);
 }
 
 export async function getuser(client: Client, user_id: string) : Promise<KResult<AccountInfo | null>> {
@@ -35,7 +35,7 @@ export async function getuser(client: Client, user_id: string) : Promise<KResult
             "SELECT user_id, pg_name, user_status, balance, status_synced, created_at, updated_at, pg_password_enc" +
             " FROM users WHERE user_id = $1",
             [user_id]
-        )
+        );
     } catch (err) {
         return Err({friendly: "User query failed"});
     }
@@ -65,8 +65,8 @@ export async function getCharges(client: Client, userId: string): Promise<Charge
         "SELECT charge_id, charge_time, user_id, charge_type, quantity, rate, amount, report_ids, transacted" +
         " FROM charges WHERE user_id = $1",
         [userId]
-    )
-    return result.rows
+    );
+    return result.rows;
 }
 
 export async function getChargesByDay(
@@ -86,7 +86,7 @@ export async function getChargesByDay(
             ORDER BY charge_time DESC
         `,
         [userId, startTime, endTime]
-    )
+    );
     return result.rows;
 }}
 
@@ -112,8 +112,8 @@ export async function getExternalTransactions(client: Client, userId: string): P
             WHERE user_id = $1
         `,
         [userId]
-    )
-    return result.rows
+    );
+    return result.rows;
 }
 
 export async function createExternalTransaction(client: Client, exttxn: NewExternalTransactionInfo): Promise<number> {
@@ -121,8 +121,8 @@ export async function createExternalTransaction(client: Client, exttxn: NewExter
         text: "SELECT new_balance FROM add_external_deposit($1, $2, $3)",
         values: [exttxn.user_id, exttxn.amount, exttxn.exttxn_extid],
         rowMode: "array",
-    })
-    return result.rows[0][0]
+    });
+    return result.rows[0][0];
 }
 
 export async function createUser(client: Client, newUser: NewUserInfo): Promise<KResult<AccountInfo>> {
@@ -158,7 +158,7 @@ function encryptPassword(password: string): KResult<Uint8Array> {
     const pgpassKeyB64 = process.env.PGPASS_KEY_B64;
     if (pgpassKeyB64 === undefined) {
         logger.error("Postgres encryption key not defined");
-        return Err({friendly: "Postgres encryption key not defined"})
+        return Err({friendly: "Postgres encryption key not defined"});
     }
     const binaryKey = Buffer.from(pgpassKeyB64, "base64");
     return Ok(encryptDataWithKey(password, binaryKey));
