@@ -2,7 +2,7 @@
 // import { Result } from "pratica"
 
 export type Result<T, E = undefined> = { isOk: true, isErr: false, value: T }
-    | { isOk: false, isErr: true, error: E }
+    | { isOk: false, isErr: true, error: E };
 
 export const Ok = <T>(data: T): Result<T, never> => {
     return { isOk: true, isErr: false, value: data };
@@ -17,4 +17,4 @@ export interface KestrelError {
     cause?: any
 }
 
-export type KResult<T> = Result<T, KestrelError>
+export type KResult<T> = Result<T, KestrelError>;
