@@ -1,8 +1,8 @@
-import pino from "pino"
+import pino from "pino";
 
 const logger = pino({
     level: "debug",
     redact: ["pg_password"],
-})
+});
 
-export default logger
+export default logger;
